Fixed lifetime unsoundness
Basically, we had a lifetime `'a` that was tied to the lifetime of the `Context` but _not_ tied to the lifetime of its parent `Fingerprint`.
So, when the `Fingerprint` was dropped, the inner type stuck around, since nothing needed to be dropped within it (just a reference).

To fix this:

1) Store raw pointers in the struct.
2) Remove the lifetime on `Raw` and `Base64` to indicate that these are owned types.
3) Add a `PhantomData` to each of the above to signal to the Rust drop checker that these structs need to be dropped.